+++ EnhanceQoL/Settings/GearUpgrade.lua
@@ -5,6 +5,7 @@ local L = LibStub("AceLocale-3.0"):GetLocale(addonName)
 local cGearUpgrade = addon.SettingsLayout.rootGENERAL
 local expandable = addon.functions.SettingsCreateExpandableSection(cGearUpgrade, {
 	name = L["GearUpgrades"],
+	newTagID = "GearUpgrades",
 	expanded = false,
 	colorizeTitle = false,
 })